fix: конвертация kotlin.Double в java.lang.Double через Double.valueOf()

+++ app/src/main/java/com/example/hotel_app/presentation/viewmodel/BookingViewModel.kt
@@ -127,7 +127,7 @@ class BookingViewModel(private val repository: HotelRepository) : ViewModel() {
                         BookingUiEvent.NavigateToPayment(
                             booking = result.booking,
                             nfcKey = result.nfcKey,
-                            amount = totalPrice
+                            amount = java.lang.Double.valueOf(totalPrice)
                         )
                     )
                 }
